Fixes after ESLint report
