+++ web/.eslintignore
@@ -1,3 +1,4 @@
 node_modules/
 .cache/*
 .netlify/*
+**/*.css
