+++ web/app/features/Bookmarks/components/Bookmarks/Bookmarks.tsx
@@ -4,9 +4,7 @@ import { Prose, Link } from '~/components';
 import { DateTimeUtils } from '~/util';
 import { Headings } from '~/components';
 import { useEffect, useState } from 'react';
-import { Form } from 'remix';
-import classNames from 'classnames';
-import { useDisabled } from '~/hooks';
+import { DeleteBookmark } from './DeleteBookmark';
 
 export interface BookmarksProps {
   bookmarks: Bookmark[];
@@ -26,8 +24,6 @@ export function Bookmarks({ bookmarks }: BookmarksProps) {
       <>
         <Headings.Content>Bookmarks</Headings.Content>
         {bookmarks.map(bookmark => {
-          const disabled = useDisabled('bookmarkId', bookmark.id);
-
           return (
             <dl key={bookmark.id} className="mb-6">
               <dt>
@@ -37,19 +33,7 @@ export function Bookmarks({ bookmarks }: BookmarksProps) {
               </dt>
               <dd>
                 Bookmarked on {DateTimeUtils.date(bookmark.createdAt)}
-                <Form action={route('/bookmarks')} method="post">
-                  <input type="hidden" name="bookmarkId" value={bookmark.id ?? ''} />
-                  <input type="hidden" name="referrer" value={referrer} />
-                  <button
-                    type="submit"
-                    className={classNames('text-sm text-red-500', {
-                      'opacity-50': disabled,
-                    })}
-                    disabled={disabled}
-                  >
-                    [delete]
-                  </button>
-                </Form>
+                <DeleteBookmark bookmark={bookmark} referrer={referrer} />
               </dd>
             </dl>
           );
